+++ classes/collectionhistory.class.php
@@ -1,18 +1,14 @@
 <?php
 
 /*
-Version:     1.1
-Date:        20/12/25
+Version:     1.2
+Date:        21/12/25
 Name:        collectionhistory.class.php
 Purpose:     Collection value history retrieval and export helpers.
 Notes:       -
 Author:      Simon Wilson
 Copyright:   2025 MTG Collection
 To do:       -
-
-History:
-    1.0 20/12/25 Initial version
-    1.1 20/12/25 Allow configurable server email for exports
 */
 
 // phpcs:disable PSR1.Classes.ClassDeclaration.MissingNamespace, PSR1.Files.SideEffects.FoundWithSymbols
@@ -167,10 +163,11 @@ class CollectionHistory
             $tempFile = tempnam(sys_get_temp_dir(), 'history_');
             file_put_contents($tempFile, $csv);
 
+            $siteTitleEsc = htmlspecialchars($this->siteTitle, ENT_QUOTES, 'UTF-8');
             $subject = "$this->siteTitle weekly value history export";
-            $emailbody = "Hi $userName, your weekly value history export from $this->siteTitle is attached."
+            $emailbody = "Hi $userName, your weekly value history export from $siteTitleEsc is attached."
                 . "<br><br>Opt out of automated emails in your profile at "
-                . "<a href='$myURL/profile.php'>your $this->siteTitle profile page</a>";
+                . "<a href='$myURL/profile.php'>your $siteTitleEsc profile page</a>";
             $emailaltbody = "Hi $userName, please see attached your weekly value history export from "
                 . "$this->siteTitle.\r\n\r\nOpt out of automated emails in your profile at your "
                 . "$this->siteTitle profile page ($myURL/profile.php)\r\n\r\n";