+++ classes/passwordcheck.class.php
@@ -1,8 +1,8 @@
 <?php
 
 /*
-Version:     2.8
-Date:        05/12/25
+Version:     2.9
+Date:        21/12/25
 Name:        passwordcheck.class.php
 Purpose:     Password validation class.
 Notes:       {none}
@@ -20,6 +20,7 @@ History:
     2.6 04/12/25 Enforce complexity and difference checks for token resets
     2.7 04/12/25 Notify user on password change
     2.8 05/12/25 Reduce reset token TTL to 10 minutes
+    2.9 21/12/25 Avoid HTML-escaping site title in email subjects and plain text
 */
 
 // phpcs:disable PSR1.Classes.ClassDeclaration.MissingNamespace, PSR1.Files.SideEffects.FoundWithSymbols
@@ -84,8 +85,7 @@ class PasswordCheck
         endif;
 
         $link = rtrim($myURL, '/') . "/reset.php?email=" . urlencode($email) . "&token=" . urlencode($token);
-        $siteTitleEsc = htmlspecialchars($siteTitle, ENT_QUOTES, 'UTF-8');
-        return $this->sendResetEmail($email, $link, $siteTitleEsc, $serverEmail, $smtpParameters);
+        return $this->sendResetEmail($email, $link, $siteTitle, $serverEmail, $smtpParameters);
     }
 
     /**
@@ -225,11 +225,10 @@ class PasswordCheck
                     if ($reset === 1) :
                         $from = "From: $serverEmail\r\nReturn-path: $serverEmail";
                         $subject = "Password reset";
-                        $siteTitleEsc = htmlspecialchars($siteTitle, ENT_QUOTES, 'UTF-8');
-                        $message = "A new password was requested for your email at $siteTitleEsc ($myURL)\n\n"
+                        $message = "A new password was requested for your email at $siteTitle ($myURL)\n\n"
                             . "Please login with this temporary password: $randompassword\n"
                             . "You will need to then choose a new password.\n\n"
-                            . "If you did not request a new password at $siteTitleEsc, you can ignore this email.";
+                            . "If you did not request a new password at $siteTitle, you can ignore this email.";
                         if ($emailEnabled) :
                             mail($email, $subject, $message, $from);
                         else :
@@ -360,13 +359,13 @@ class PasswordCheck
         endif;
 
         $siteTitleEsc = htmlspecialchars($siteTitle, ENT_QUOTES, 'UTF-8');
-        $subject = "$siteTitleEsc password changed";
-        $plain = "Your password on $siteTitleEsc was changed. "
+        $subject = "$siteTitle password changed";
+        $plain = "Your password on $siteTitle was changed. "
                   . "If this was not you, please reset your password immediately.";
         $html = "<p>Your password on $siteTitleEsc was changed.</p>"
               . "<p>If this was not you, please reset your password immediately.</p>";
 
-        $mailer = new MyPHPMailer(true, $smtpParameters, $serverEmail, $this->logfile, $siteTitleEsc);
+        $mailer = new MyPHPMailer(true, $smtpParameters, $serverEmail, $this->logfile, $siteTitle);
         if ($mailer->sendEmail($email, true, $subject, $html, $plain)) :
             $this->message->logMessage('[NOTICE]', "Password change notification sent to $email");
             return true;
@@ -505,9 +504,8 @@ class PasswordCheck
             return false;
         endif;
 
-        $siteTitleEsc = htmlspecialchars($siteTitle, ENT_QUOTES, 'UTF-8');
-        $mail = new MyPHPMailer(true, $smtpParameters, $serverEmail, $this->logfile, $siteTitleEsc);
-        $subject = "$siteTitleEsc password reset";
+        $mail = new MyPHPMailer(true, $smtpParameters, $serverEmail, $this->logfile, $siteTitle);
+        $subject = "$siteTitle password reset";
         $safeLink = htmlspecialchars($link, ENT_QUOTES, 'UTF-8');
         $bodyText = "A password reset was requested for your account. Click the link below to set a new password:\n\n"
             . "$link\n\nIf you did not request this, you can ignore this email.";